Updated the way we save scalers

+++ src/io/file/hl_file_io.h
@@ -94,9 +94,9 @@ namespace io
 
 struct DLL_EXPORT file_name_size
 {
-    file_name_size(std::string name, int size) { filename = name; total_rank_size = size;}
+    file_name_size(std::string name, long size) { filename = name; total_rank_size = size;}
     std::string filename;
-    int total_rank_size;
+    long total_rank_size;
 };
 
 struct DLL_EXPORT file_name_fit_params
